# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "CompareSignals": {
        "additionalProperties": false,
        "properties": {
          "confidence_drop": {
            "default": 20,
            "exclusiveMinimum": 0.0,
            "maximum": 100.0,
            "title": "Confidence Drop",
            "type": "number"
          },
          "rank_change": {
            "default": 5,
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Rank Change",
            "type": "integer"
          },
          "score_change": {
            "default": 10,
            "exclusiveMinimum": 0.0,
            "maximum": 100.0,
            "title": "Score Change",
            "type": "number"
          },
          "snapshot_id": {
            "minimum": 1.0,
            "title": "Snapshot Id",
            "type": "integer"
          }
        },
        "required": [
          "snapshot_id"
        ],
        "title": "CompareSignals",
        "type": "object"
      },
...
      "EarningsEvent": {
        "additionalProperties": false,
        "properties": {
          "days_until": {
            "title": "Days Until",
            "type": "integer"
          },
          "event_date": {
            "format": "date",
            "title": "Event Date",
            "type": "string"
          },
          "is_estimate": {
            "title": "Is Estimate",
            "type": "boolean"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          }
        },
        "required": [
          "symbol",
          "event_date",
          "is_estimate",
          "days_until"
        ],
        "title": "EarningsEvent",
        "type": "object"
      },
      "EarningsList": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EarningsEvent"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "EarningsList",
        "type": "object"
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "status": {
            "default": "error",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "FilingCheckSaved": {
        "additionalProperties": false,
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/SignalEvent"
            },
            "title": "Events",
            "type": "array"
          },
          "recorded": {
            "title": "Recorded",
            "type": "boolean"
          }
        },
        "required": [
          "recorded",
          "events"
        ],
        "title": "FilingCheckSaved",
        "type": "object"
      },
...
      "SaveFilingCheck": {
        "additionalProperties": false,
        "properties": {
          "job_id": {
            "pattern": "^[a-f0-9]{32}$",
            "title": "Job Id",
            "type": "string"
          }
        },
        "required": [
          "job_id"
        ],
        "title": "SaveFilingCheck",
        "type": "object"
      },
      "SaveSnapshot": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "top_n": {
            "default": 10,
            "maximum": 30.0,
            "minimum": 1.0,
            "title": "Top N",
            "type": "integer"
          }
        },
        "required": [
          "name"
        ],
        "title": "SaveSnapshot",
        "type": "object"
      },
...
      "SignalEvent": {
        "additionalProperties": false,
        "properties": {
          "cause": {
            "title": "Cause",
            "type": "string"
          },
          "detected_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detected At"
          },
          "event_type": {
            "title": "Event Type",
            "type": "string"
          },
          "from_snapshot_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "From Snapshot Id"
          },
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "new_value": {
            "title": "New Value"
          },
          "previous_value": {
            "title": "Previous Value"
          },
          "severity": {
            "title": "Severity",
            "type": "string"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "to_snapshot_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "To Snapshot Id"
          }
        },
        "required": [
          "symbol",
          "event_type",
          "severity",
          "cause"
        ],
        "title": "SignalEvent",
        "type": "object"
      },
      "SignalEventList": {
        "additionalProperties": false,
        "properties": {
          "basis": {
            "const": "frozen_weights",
            "default": "frozen_weights",
            "title": "Basis",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/SignalEvent"
            },
            "title": "Items",
            "type": "array"
          },
          "note": {
            "default": "Son cambios de datos respecto a un snapshot, no órdenes ni prueba de ventaja sobre SPY.",
            "title": "Note",
            "type": "string"
          },
          "status": {
            "default": "DIAGNOSTIC",
            "enum": [
              "DIAGNOSTIC",
              "EXPERIMENTAL"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "items"
        ],
        "title": "SignalEventList",
        "type": "object"
      },
...
      "Snapshot": {
        "additionalProperties": false,
        "properties": {
          "as_of_date": {
            "format": "date",
            "title": "As Of Date",
            "type": "string"
          },
          "candidates": {
            "title": "Candidates",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "created_at",
          "as_of_date",
          "source",
          "candidates"
        ],
        "title": "Snapshot",
        "type": "object"
      },
      "SnapshotList": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Snapshot"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "SnapshotList",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "GABI local API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/market/signals": {
      "get": {
        "operationId": "events_api_v1_market_signals_get",
        "parameters": [
          {
            "in": "query",
            "name": "severity",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "INFO",
                    "WATCH",
                    "MATERIAL"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignalEventList"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Modo no permitido"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Empresa no encontrada"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Datos cambiados"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Consulta no válida"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error interno"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Datos no disponibles"
          }
        },
        "summary": "Events",
        "tags": [
          "market"
        ]
      }
    },
    "/api/v1/market/signals/compare": {
      "post": {
        "operationId": "compare_api_v1_market_signals_compare_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompareSignals"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignalEventList"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Modo no permitido"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Empresa no encontrada"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Datos cambiados"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Consulta no válida"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error interno"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Datos no disponibles"
          }
        },
        "summary": "Compare",
        "tags": [
          "market"
        ]
      }
    },
    "/api/v1/market/signals/filings/record": {
      "post": {
        "operationId": "record_filings_api_v1_market_signals_filings_record_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveFilingCheck"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilingCheckSaved"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Modo no permitido"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Empresa no encontrada"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Datos cambiados"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Consulta no válida"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error interno"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Datos no disponibles"
          }
        },
        "summary": "Record Filings",
        "tags": [
          "market"
        ]
      }
    },
    "/api/v1/market/snapshots": {
      "get": {
        "operationId": "snapshots_api_v1_market_snapshots_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotList"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Modo no permitido"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Empresa no encontrada"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Datos cambiados"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Consulta no válida"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error interno"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Datos no disponibles"
          }
        },
        "summary": "Snapshots",
        "tags": [
          "market"
        ]
      },
      "post": {
        "operationId": "save_snapshot_api_v1_market_snapshots_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveSnapshot"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Snapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Modo no permitido"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Empresa no encontrada"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Datos cambiados"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Consulta no válida"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error interno"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Datos no disponibles"
          }
        },
        "summary": "Save Snapshot",
        "tags": [
          "market"
        ]
      }
    },
    "/api/v1/market/snapshots/{snapshot_id}/earnings": {
      "get": {
        "operationId": "earnings_api_v1_market_snapshots__snapshot_id__earnings_get",
        "parameters": [
          {
            "in": "path",
            "name": "snapshot_id",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Snapshot Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EarningsList"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Modo no permitido"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Empresa no encontrada"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Datos cambiados"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Consulta no válida"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error interno"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Datos no disponibles"
          }
        },
        "summary": "Earnings",
        "tags": [
          "market"
        ]
      }
    },
...
  }
}